move gpu metrics render logic into gpu metric parser
